+++ OpenDreamRuntime/Procs/Native/DreamProcNativeRoot.cs
@@ -1014,64 +1014,9 @@ internal static class DreamProcNativeRoot {
     [DreamProcParameter("Depth", Type = DreamValueTypeFlag.Float)]
     [DreamProcParameter("Center", Type = DreamValueTypeFlag.DreamObject)]
     public static DreamValue NativeProc_hearers(NativeProc.Bundle bundle, DreamObject? src, DreamObject? usr) { //TODO: Change depending on center
-        DreamValue depthValue = new DreamValue(5);
-        DreamObjectAtom? center = null;
-
-        //Arguments are optional and can be passed in any order
-        if (bundle.Arguments.Length > 0) {
-            DreamValue firstArgument = bundle.GetArgument(0, "Depth");
-
-            if (firstArgument.TryGetValueAsDreamObject(out center)) {
-                if (bundle.Arguments.Length > 1) {
-                    depthValue = bundle.GetArgument(1, "Center");
-                }
-            } else {
-                depthValue = firstArgument;
-
-                if (bundle.Arguments.Length > 1) {
-                    bundle.GetArgument(1, "Center").TryGetValueAsDreamObject(out center);
-                }
-            }
-        }
-
-        center ??= usr as DreamObjectAtom;
-
-        DreamList hear = bundle.ObjectTree.CreateList();
-        if (center == null)
-            return new(hear);
-
-        var centerPos = bundle.AtomManager.GetAtomPosition(center);
-        if (!depthValue.TryGetValueAsInteger(out var depth))
-            depth = 5;
-
-        foreach (var atom in bundle.AtomManager.EnumerateAtoms(bundle.ObjectTree.Mob)) {
-            var mob = (DreamObjectMob)atom;
-
-            if (centerPos.Z == mob.Z && Math.Abs(centerPos.X - mob.X) <= depth && Math.Abs(centerPos.Y - mob.Y) <= depth) {
-                (_, ViewRange range) = DreamProcNativeHelpers.ResolveViewArguments(bundle.DreamManager, mob, bundle.Arguments);
-                var earPos = bundle.AtomManager.GetAtomPosition(mob);
-                var viewData = DreamProcNativeHelpers.CollectViewData(bundle.AtomManager, bundle.MapManager, earPos, range);
-
-                ViewAlgorithm.CalculateVisibility(viewData, true);
-
-                for (int x = 0; x < viewData.GetLength(0); x++) {
-                    for (int y = 0; y < viewData.GetLength(1); y++) {
-                        var tile = viewData[x, y];
-                        if (tile == null || tile.IsVisible == false)
-                            continue;
-
-                        if (centerPos.X == earPos.X + tile.DeltaX && earPos.Y + tile.DeltaY == centerPos.Y) {
-                            hear.AddValue(new DreamValue(mob));
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        return new DreamValue(hear);
+        return DreamProcNativeHelpers.HandleViewersHearers(bundle, usr, true);
     }
-  
+
     [DreamProc("html_decode")]
     [DreamProcParameter("HtmlText", Type = DreamValueTypeFlag.String)]
     public static DreamValue NativeProc_html_decode(NativeProc.Bundle bundle, DreamObject? src, DreamObject? usr) {
@@ -1863,7 +1808,7 @@ internal static class DreamProcNativeRoot {
             }
             return new DreamValue(floatNum.ToString($"g{sigFig}"));
         }
-        
+
         if(bundle.Arguments.Length == 3) {
             var digits = Math.Max(bundle.GetArgument(1, "A").MustGetValueAsInteger(), 1);
             var radix = bundle.GetArgument(2, "B").MustGetValueAsInteger();
@@ -1875,70 +1820,12 @@ internal static class DreamProcNativeRoot {
         // Maybe an exception is better?
         return new DreamValue("0");
     }
- 
+
     [DreamProc("ohearers")]
     [DreamProcParameter("Depth", Type = DreamValueTypeFlag.Float)]
     [DreamProcParameter("Center", Type = DreamValueTypeFlag.DreamObject)]
     public static DreamValue NativeProc_ohearers(NativeProc.Bundle bundle, DreamObject? src, DreamObject? usr) { //TODO: Change depending on center
-        DreamValue depthValue = new DreamValue(5);
-        DreamObjectAtom? center = null;
-
-        //Arguments are optional and can be passed in any order
-        if (bundle.Arguments.Length > 0) {
-            DreamValue firstArgument = bundle.GetArgument(0, "Depth");
-
-            if (firstArgument.TryGetValueAsDreamObject(out center)) {
-                if (bundle.Arguments.Length > 1) {
-                    depthValue = bundle.GetArgument(1, "Center");
-                }
-            } else {
-                depthValue = firstArgument;
-
-                if (bundle.Arguments.Length > 1) {
-                    bundle.GetArgument(1, "Center").TryGetValueAsDreamObject(out center);
-                }
-            }
-        }
-
-        center ??= usr as DreamObjectAtom;
-
-        DreamList hear = bundle.ObjectTree.CreateList();
-        if (center == null)
-            return new(hear);
-
-        var centerPos = bundle.AtomManager.GetAtomPosition(center);
-        if (!depthValue.TryGetValueAsInteger(out var depth))
-            depth = 5; //TODO: Default to world.view
-
-        foreach (var atom in bundle.AtomManager.EnumerateAtoms(bundle.ObjectTree.Mob)) {
-            var mob = (DreamObjectMob)atom;
-
-            if (mob.X == centerPos.X && mob.Y == centerPos.Y)
-                continue;
-
-            if (centerPos.Z == mob.Z && Math.Abs(centerPos.X - mob.X) <= depth && Math.Abs(centerPos.Y - mob.Y) <= depth) {
-                (_, ViewRange range) = DreamProcNativeHelpers.ResolveViewArguments(bundle.DreamManager, mob, bundle.Arguments);
-                var earPos = bundle.AtomManager.GetAtomPosition(mob);
-                var viewData = DreamProcNativeHelpers.CollectViewData(bundle.AtomManager, bundle.MapManager, earPos, range);
-
-                ViewAlgorithm.CalculateVisibility(viewData, true);
-
-                for (int x = 0; x < viewData.GetLength(0); x++) {
-                    for (int y = 0; y < viewData.GetLength(1); y++) {
-                        var tile = viewData[x, y];
-                        if (tile == null || tile.IsVisible == false)
-                            continue;
-
-                        if (centerPos.X == earPos.X + tile.DeltaX && earPos.Y + tile.DeltaY == centerPos.Y) {
-                            hear.AddValue(new DreamValue(mob));
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        return new DreamValue(hear);
+        return DreamProcNativeHelpers.HandleOviewersOhearers(bundle, usr, true);
     }
 
     [DreamProc("orange")]
@@ -1993,65 +1880,7 @@ internal static class DreamProcNativeRoot {
     [DreamProcParameter("Depth", Type = DreamValueTypeFlag.Float)]
     [DreamProcParameter("Center", Type = DreamValueTypeFlag.DreamObject)]
     public static DreamValue NativeProc_oviewers(NativeProc.Bundle bundle, DreamObject? src, DreamObject? usr) { //TODO: Change depending on center
-        DreamValue depthValue = new DreamValue(5);
-        DreamObjectAtom? center = null;
-
-        //Arguments are optional and can be passed in any order
-        if (bundle.Arguments.Length > 0) {
-            DreamValue firstArgument = bundle.GetArgument(0, "Depth");
-
-            if (firstArgument.TryGetValueAsDreamObject(out center)) {
-                if (bundle.Arguments.Length > 1) {
-                    depthValue = bundle.GetArgument(1, "Center");
-                }
-            } else {
-                depthValue = firstArgument;
-
-                if (bundle.Arguments.Length > 1) {
-                    bundle.GetArgument(1, "Center").TryGetValueAsDreamObject(out center);
-                }
-            }
-        }
-
-        center ??= usr as DreamObjectAtom;
-
-        DreamList view = bundle.ObjectTree.CreateList();
-        if (center == null)
-            return new(view);
-
-        var centerPos = bundle.AtomManager.GetAtomPosition(center);
-        if (!depthValue.TryGetValueAsInteger(out var depth))
-            depth = 5; //TODO: Default to world.view
-
-        foreach (var atom in bundle.AtomManager.EnumerateAtoms(bundle.ObjectTree.Mob)) {
-            var mob = (DreamObjectMob)atom;
-
-            if (mob.X == centerPos.X && mob.Y == centerPos.Y)
-                continue;
-
-            if (centerPos.Z == mob.Z && Math.Abs(centerPos.X - mob.X) <= depth && Math.Abs(centerPos.Y - mob.Y) <= depth) {
-                (_, ViewRange range) = DreamProcNativeHelpers.ResolveViewArguments(bundle.DreamManager, mob, bundle.Arguments);
-                var eyePos = bundle.AtomManager.GetAtomPosition(mob);
-                var viewData = DreamProcNativeHelpers.CollectViewData(bundle.AtomManager, bundle.MapManager, eyePos, range);
-
-                ViewAlgorithm.CalculateVisibility(viewData);
-
-                for (int x = 0; x < viewData.GetLength(0); x++) {
-                    for (int y = 0; y < viewData.GetLength(1); y++) {
-                        var tile = viewData[x, y];
-                        if (tile == null || tile.IsVisible == false)
-                            continue;
-
-                        if (centerPos.X == eyePos.X + tile.DeltaX && eyePos.Y + tile.DeltaY == centerPos.Y) {
-                            view.AddValue(new DreamValue(mob));
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        return new DreamValue(view);
+        return DreamProcNativeHelpers.HandleOviewersOhearers(bundle, usr, false);
     }
 
     public static string list2params(DreamList list) {
@@ -3218,62 +3047,7 @@ internal static class DreamProcNativeRoot {
     [DreamProcParameter("Depth", Type = DreamValueTypeFlag.Float)]
     [DreamProcParameter("Center", Type = DreamValueTypeFlag.DreamObject)]
     public static DreamValue NativeProc_viewers(NativeProc.Bundle bundle, DreamObject? src, DreamObject? usr) { //TODO: Change depending on center
-        DreamValue depthValue = new DreamValue(5);
-        DreamObjectAtom? center = null;
-
-        //Arguments are optional and can be passed in any order
-        if (bundle.Arguments.Length > 0) {
-            DreamValue firstArgument = bundle.GetArgument(0, "Depth");
-
-            if (firstArgument.TryGetValueAsDreamObject(out center)) {
-                if (bundle.Arguments.Length > 1) {
-                    depthValue = bundle.GetArgument(1, "Center");
-                }
-            } else {
-                depthValue = firstArgument;
-
-                if (bundle.Arguments.Length > 1) {
-                    bundle.GetArgument(1, "Center").TryGetValueAsDreamObject(out center);
-                }
-            }
-        }
-
-        center ??= usr as DreamObjectAtom;
-
-        DreamList view = bundle.ObjectTree.CreateList();
-        if (center == null)
-            return new(view);
-
-        var centerPos = bundle.AtomManager.GetAtomPosition(center);
-        if (!depthValue.TryGetValueAsInteger(out var depth))
-            depth = 5; //TODO: Default to world.view
-
-        foreach (var atom in bundle.AtomManager.EnumerateAtoms(bundle.ObjectTree.Mob)) {
-            var mob = (DreamObjectMob)atom;
-
-            if (centerPos.Z == mob.Z && Math.Abs(centerPos.X - mob.X) <= depth && Math.Abs(centerPos.Y - mob.Y) <= depth) {
-                (_, ViewRange range) = DreamProcNativeHelpers.ResolveViewArguments(bundle.DreamManager, mob, bundle.Arguments);
-                var eyePos = bundle.AtomManager.GetAtomPosition(mob);
-                var viewData = DreamProcNativeHelpers.CollectViewData(bundle.AtomManager, bundle.MapManager, eyePos, range);
-
-                ViewAlgorithm.CalculateVisibility(viewData);
-
-                for (int x = 0; x < viewData.GetLength(0); x++) {
-                    for (int y = 0; y < viewData.GetLength(1); y++) {
-                        var tile = viewData[x, y];
-                        if (tile == null || tile.IsVisible == false)
-                            continue;
-
-                        if (centerPos.X == eyePos.X + tile.DeltaX && eyePos.Y + tile.DeltaY == centerPos.Y) {
-                            view.AddValue(new DreamValue(mob));
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        return new DreamValue(view);
+        return DreamProcNativeHelpers.HandleViewersHearers(bundle, usr, false);
     }
 
     [DreamProc("walk")]
